+++ src/Manager/ShipTerminal/TerminalSubCommand.cs
@@ -3,22 +3,22 @@ using System.Collections.Generic;
 
 namespace QualityCompany.Manager.ShipTerminal;
 
-public class TerminalSubCommand
+internal class TerminalSubCommand
 {
-    public string Id;
-    public string Name;
-    public string Message;
-    public TerminalKeyword Keyword;
-    public TerminalNode Node;
-    public List<string> Conditions = new();
-    public Action PreConditionAction;
-    public Func<string, bool> VariablePreAction;
-    public string ActionEvent;
-    public Action Action;
+    internal string Id;
+    internal string Name;
+    internal string Message;
+    internal TerminalKeyword Keyword;
+    internal TerminalNode Node;
+    internal List<string> Conditions = new();
+    internal Action PreConditionAction;
+    internal Func<string, bool> VariablePreAction;
+    internal string ActionEvent;
+    internal Action Action;
 
-    public bool IsVariableCommand;
-    public string VariableRegexMatchPattern;
+    internal bool IsVariableCommand;
+    internal string VariableRegexMatchPattern;
 
-    public string ConfirmMessage = "Confirmed!";
-    public string DenyMessage = "Cancelled!";
+    internal string ConfirmMessage = "Confirmed!";
+    internal string DenyMessage = "Cancelled!";
 }