+++ src/Modules/Ship/TimeMonitor.cs
@@ -14,9 +14,9 @@ internal class TimeMonitor : BaseMonitor
     {
         Instance = this;
 
-        _logger = new ACLogger(nameof(TimeMonitor));
+        Logger = new ACLogger(nameof(TimeMonitor));
         _timeMesh = GameObject.Find("Systems/UI/Canvas/IngamePlayerHUD/ProfitQuota/Container/Box/TimeNumber").GetComponent<TextMeshProUGUI>();
-        _textMesh.text = "TIME:\n7:30\nAM";
+        TextMesh.text = "TIME:\n7:30\nAM";
     }
 
     private void OnDestroy()