+++ src/Network/NetworkHandler.cs
@@ -20,12 +20,6 @@ internal class NetworkHandler : NetworkBehaviour
     internal void UpdateSellTargetClientRpc(int newTarget, string playerName)
     {
         TargetManager.UpdateTargetClient(newTarget, playerName);
-
-        CompanyNetworkHandler.Instance.SaveData.TargetForSelling = newTarget;
-        if (IsHost)
-        {
-            CompanyNetworkHandler.Instance.ServerSaveFileServerRpc();
-        }
     }
 
     [ServerRpc(RequireOwnership = false)]
@@ -66,13 +60,6 @@ internal class NetworkHandler : NetworkBehaviour
 
     public override void OnNetworkSpawn()
     {
-        if (NetworkManager.Singleton.IsHost || NetworkManager.Singleton.IsServer)
-        {
-            // This is recommended in the docs, but it doesn't seem to work
-            // https://lethal.wiki/dev/advanced/networking#preventing-duplication
-            // Instance?.gameObject.GetComponent<NetworkObject>().Despawn();
-        }
-
         Instance = this;
 
         base.OnNetworkSpawn();