+++ src/Patch/GameNetworkManagerPatch.cs
@@ -1,5 +1,5 @@
 ﻿using HarmonyLib;
-using QualityCompany.Network;
+using QualityCompany.Manager.Saves;
 using QualityCompany.Utils;
 using static QualityCompany.Service.GameEvents;
 
@@ -12,18 +12,15 @@ internal class GameNetworkManagerPatch
     [HarmonyPatch("Disconnect")]
     private static void DisconnectPatch(GameNetworkManager __instance)
     {
-        GameUtils.Reset();
-
         OnDisconnected(__instance);
+
+        GameUtils.Reset();
     }
 
     [HarmonyPrefix]
     [HarmonyPatch("SaveGame")]
-    private static void SaveGamePatch(GameNetworkManager __instance)
+    private static void SaveGamePatch()
     {
-        if (!Plugin.Instance.PluginConfig.NetworkingEnabled) return;
-        if (!__instance.isHostingGame) return;
-
-        CompanyNetworkHandler.Instance.ServerSaveFileServerRpc();
+        SaveManager.Save();
     }
 }